// src/app/services/generated/model/sensorType.ts
/**
 * Read management system Frontend API
 * Read management system Frontend API
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export type SensorType = 'traffic';

export const SensorType = {
    Traffic: 'traffic' as SensorType
};