// src/app/services/generated/model/voivodeship.ts
/**
 * Read management system Frontend API
 * Read management system Frontend API
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export type Voivodeship = 'dolnoslaskie' | 'kujawskopomorskie' | 'lubelskie' | 'lubuskie' | 'lodzkie' | 'malopolskie' | 'mazowieckie' | 'opolskie' | 'podkarpackie' | 'podlaskie' | 'pomorskie' | 'slaskie' | 'swietokrzyskie' | 'warminskomazurskie' | 'wielkopolskie' | 'zachodniopomorskie';

export const Voivodeship = {
    Dolnoslaskie: 'dolnoslaskie' as Voivodeship,
    Kujawskopomorskie: 'kujawskopomorskie' as Voivodeship,
    Lubelskie: 'lubelskie' as Voivodeship,
    Lubuskie: 'lubuskie' as Voivodeship,
    Lodzkie: 'lodzkie' as Voivodeship,
    Malopolskie: 'malopolskie' as Voivodeship,
    Mazowieckie: 'mazowieckie' as Voivodeship,
    Opolskie: 'opolskie' as Voivodeship,
    Podkarpackie: 'podkarpackie' as Voivodeship,
    Podlaskie: 'podlaskie' as Voivodeship,
    Pomorskie: 'pomorskie' as Voivodeship,
    Slaskie: 'slaskie' as Voivodeship,
    Swietokrzyskie: 'swietokrzyskie' as Voivodeship,
    Warminskomazurskie: 'warminskomazurskie' as Voivodeship,
    Wielkopolskie: 'wielkopolskie' as Voivodeship,
    Zachodniopomorskie: 'zachodniopomorskie' as Voivodeship
};